feat(localFileSearch): restructure LocalFileSearchTab component and update import paths
